requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.3.0
Django==4.2.21
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
psycopg==3.2.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.7.0
